Handle delay slots during the branch-delays that cause them
In theory this commit should boost performance, but in practice it seems to
make it worse.  I think this is because the delay slots are currently
"free cycles" (they would not be free in real life) which degrades performance
by overclocking the guest CPU.  The solution is to implement cycle counting for
the delay slots.

One other downside is that the debugger shouldn't be able to break on delay
slots anymore.  That's unfortunate, but if I can get this commit to boost
performance instead of degrade it then this sacrifice it will be worth it.

Also, I think this breaks the system call tracking.  TBH I don't use it
anyways, the only time I've ever used it was when I was trying to figure
out why the firmware was rejecting my gdrom drive and that was way back in
August.