revert commit de8e4a84b016b75207241767d2100ac4b159fc49
The purpose of reverting this commit is to make way for the dynamic recompiler.
This commit was good for performance (and removing it is therefore bad for
performance), but it would interfere with the jit since I wouldn't be able to
calculate cycle-counts outside of the instruction handlers.

One alternative would have been to remove the cycle-counting from the
instruction handlers and then wrap them with handlers that call the
non-cycle-count versions of the handlers, and then do the cycle-counting.  That
would work, but then I would have to add the function pointers to Sh4InstOpcode,
and also maintain more code.