basic JIT implementation
The idea for the initial implementation is to to create something that is
essentially a threaded interpreter, except it will have dedicated branch-delay
implementations since implementing branch delay slots in a pure threaded
interpreter seems very difficult to do (and also counter-productive since I
intend for this to evolve into a fully dynamic recompiler).

Currently there's no emitter backend, and instead the JIT'd
intermediate-language is interpreted by an interpreter.  It is significantly
slower than the default interpreter; this is undoubtedly due to the way that I
cache code in a linear array instead of using a more efficient tree or
hashmap.