get rid of the gfx_thread entirely
The reason why this has existed as a separate thread until now is that I wanted
to implement a software renderer, and a software renderer would work best if it
had its own thread.  I stil want to have a software renderer, but the day when
I do that is still a long way off into the future, and right now I'm not even
bothering to test with separate rendering threads enabled, so I don't get much
out of this infrastructure and it ultimately serves only to make code more
complicated than it needs to be.