gfx: add a new command to take screenshots
The way I've implemented it is a little suboptimal.  Instead of copying the
framebuffer back from the GPU when the user requests a screenshot, I make a copy
of the framebuffer on the CPU-side every time a new framebuffer is posted.
Then it makes another copy when it's actually saving the PNG file.  I don't
expect this to impact performance in a noticeable way, but if it does I can
always go back and change it later.