jit: transform the jit from a threaded interpreter into a true JIT recompiler.
This is a bit of an uber-commit, but the primary change here is that the jit
now operates on "slots" instead of SH4 registers.  A slot is like a variable
for the IL language; it can refer to an SH4 register, or just some intermediate
value.  The JIT frontend (sh4_disas.c) manages the slots and maps them to SH4
registers.  The JIT backend (code_block_x86_64.c) maps the slots to x86_64
registers and the stack.  The JIT backend has no concept of what a slot
represents; all of the legwork for keeping slots in sync with the SH4 registers
they represent is done by the frontend code in sh4_disas.c