make cmake download libevent during build
I need EVLOOP_NO_EXIT_ON_EMPTY, which appears to have been added in one of the
more recent libevent releases.  There are several linux distributions which do
not keep their package managers' versions of libevent as up-to-date as I would
like; this makes it difficult to build WashingtonDC on those distributions.

To resolve this, I'm using cmake's ExternalProject module to download and build
a new copy of libevent 2.1.8 for every build.