fix -DENABLE_DEBUGGER=On builds
Unfortunately, the gdb-stub is still broken for now due to the commits I made a
few months ago that caused the memory to raise real errors instead of pending
errors when gdb pokes at a bad memory address.  I don't regret getting rid of
this infra because it was making things more difficult than they needed to be,
but I do need to come up with some sort of a side-channel that can fail so I can
bring the gdb stub back.