jit: exec_mem: fix incorrect INVARIANTS check in exec_mem_grow
I *think* that check was intended to detect overlap caused by memory corruption.
TBH, I'm not sure I remember why that's there.  All I can say for sure is that
it was causing the allocation to fail with an ERROR_INTEGRITY when it detects a
free chunk in front of the allocation being grown, and that is not something
that should be raising an ERROR_INTEGRITY because that's just normal
fragmentation.