gfx: replace geo_buf with the new gfx_il system
The idea of gfx_il is that the pvr2 will write a simple program that it sends
to the gfx code via an interpreted bytecode.  This is easier than passing
around large structures like geo_buf.  For now, geo_buf still exists, but only
in the pvr2 code.  Eventually I will remove it completely.