source glfw from a submodule
After I fixed issue #36 on github, I ran into some other build issues on
Manjaro Linux pertaining to the pkg-config path.  This was the easiest way
to get around that issue, and I've been meaning to do this for a while anyways
for the sake of the Windows port that I'm going to do eventually.