update glfw to the latest version of master
This resolves issue #37 on github.  Previously I was on tags/3.2.1, but I
encountered a bug in which WashingtonDC would segfault on closing.  This bug
was found to be a bug in glfw, not WashingtonDC (see issue #37 for proof).
Since the latest version of glfw's master branch works, I'm going to assume
that whatever bug was causing this has already been fixed upstream.

This also lets me try out the SDL_gamecontrollerdb stuff, and in general makene
sure that WashingtonDC is ready for when version 3.3 gets officially released.