commit copies of libpng and zlib into WashingtonDC's source tree
The purpose of this is to facilitate the Microsoft Windows port.  Previously
WashingtonDC would go along with whatever versions of these libraries happened
to be on the user's PC at build-time.  That works fine on Unix, but Windows
doesn't have a default directory or even an environment variable the compiler
can use to look for dependencies.

This commit also adds a license section to the README.md because I have to be
careful about that now that I'm distributing other peoples' code alongside
WashingtonDC.