jit: separate the determinism stuff into a separate pass.
Also make it possible to disable it at compile-time and disable it by default.
I never really did anything with this and I'm not even sure if it works.  I
might even delete it soon and start over from scratch in the future when I
actually have a use for this sort of thing.