cmake will now get libevent from github rather than the tarball on their website
The reason for this is that the github version of the source has cmake support,
while the tarball version of the source only has GNU autotools.  Since I want
to be able to build on Windows without the aid of cygwin or msys, autotools is
out of the question.

I also tried turning it into a submodule and building it via add_subdirectory
like I already do with glfw, but I ran into some weird build issues; it doesn't
look like their CMakeLists.txt is designed to be embedded within another
project like that.