Make sure cmake knows libevent is a cmake project
the point of this is so that it knows that the build command on Windows is
not make.  This commit also sets options to disable some things we don't
need.

With this commit, all of WashingtonDC's dependencies can build on Windows
with Visual Studio, so I can start working on actually porting
WashingtonDC itself to Windows.