stringlib: string_get_col now ignores delimiters in columns bound by quotes
This is needed because the gdi-parsing code needs to allow track-paths with
spaces in them as long as those track-paths are bound by quotes because that's
how reicast parses gdi files.

I feel comfortable fixing this in stringlib.c instead of gdi.c because gdi.c is
the only user of this function.

This resolves issue #39 on github.