jit: x86_64: generate dispatch code at runtime
The idea is that the CPU never has to leave the executable memory we mmap in
ourselves between events, so it runs faster.  My experience in the past has been
that CPUs typically run faster when they don't have to jump between different
mmap allocations.

The current implementation is suboptimal and actually serves to make
WashingtonDC slower instead of faster, but I think it will be much better once
I've gotten all the C code rewritten as JIT'd assembly.