gfx: implement gfx_obj, a gfx-side data-store infrastructure
The idea is that a gfx_obj represents some object which resides within the gfx
system's domain but is ultimately managed by the pvr2 emulation code.  A gfx_obj
can be bound to a texture or it can be bound to a render-target, and it takes
the data it stores within it when it gets rebound.  Currently it's only used by
the texture cache.  The render-target stuff will come later.