gfx: move the framebuffer over to the new gfx_obj infrastructure
Currently this infra serves only to make things slower by adding a second layer
of indirection+copying, but the eventual goal is that I'm going to be able to
implement things like render-to-texture without needlessly moving data back and
forth between OpenGL and WashingtonDC and then it's going to be a lot faster.
Also there are probably a few places where I'm doing redundant copy operations
that can be eliminated.