pvr2: framebuffer: framebuffer management
This commit can cache framebuffers in OpenGL without ever reading them back;
this helps performance bigly and gets us up to 120% average performance during
spiral swirl (relative to actual Dreamcast).

In-game in Crazy Taxi, the framerate is lower, but the avereage does appear to
be in the 90% range, so WashingtonDC is *almost* running that game at full-speed
on Bulldozer.

There are still some bugs, such as:
* The SEGA logo in IP.BIN flashes black every other frame
* need to take linestride, pixel format, interlacing, etc into account for
  framebuffer caching in hw/pvr2/framebuffer.c
* I don't ever actually write the framebuffers back to memory, which breaks
  render-to-texture (Crazy Taxi pause menu).