pvr2: spg: double the pixel clock when interlacing is enabled.
Based on my own experiments on Dreamcast, turning on interlacing does not
effect the behavior of the scanline position; it still goes from 0 to vcount-1
in increments of one.  I don't know how exactly the SPG represents interlaced
scan, but the basic idea is that it needs to set off a vblank interrupt between
fields (which is twice per frame if we define the word "frame" to include both
fields) so doubling the pixel clock has the right effect.  With this commit I
get a perfect 59.94005994 Hz, which is exactly the frequency of interlaced NTSC
composite video.

Someday I'd like to figure out exactly how vblank interrupts are supposed to
work on interlaced video.  One experiment to consider in the future is to write
a test program that will output the scanline position on each vblank interrupt.

See issue #43 on github.