MemoryMap: don't hardocde memory regions
This will allow memory read/write operations to be integrated with the JIT and
it will (hopefully) give me a chance to fix the gdb stub without impacting
performance.

Currently this commit actually makes WashingtonDC slower, but not as slow as I
had feared it might.  When memory access is integrated into the JIT, I expect
that WashingtonDC will be much faster.