MemoryMap: integrate sh4's p4 area into the memory map.
With this, there's no more need for sh4_read_mem and sh4_write_mem to exist as
separate functions from the memory map.  Once again, this commit serves only to
make WashingtonDC slower, but I believe that this will make it easier to
integrate memory access with the JIT, and once that happens WashingtonDC will be
faster than it has ever been before because the penalties of defining memory
regions at runtime will go away.