1555 ARGB pixel format
The graphics hardware can output this format when rendering, but the video
hardware doesn't support it for output.  Seems all it's good for is rendering
to a texture.

Twinkle Star Sprites uses this, although I'm not sure what for (this game
doesn't seem like it ought to need to render to texture).