debugger: fix gdb_stub by adding in MemoryMap functions that can fail gracefully
This will still cause the emulator to crash if gdb tries to access an address
that evaluates to a valid memory region, but not to a valid address within that
region.  To fix that, I need to have special try_read/try_write handlers for
each memory map region as well.

This commit resolves issue #35 on github.