support for multiple clocks.
Right now there's only one in use, but the eventual goal is to have mroe than
one so the ARM7 and the SH4 can be running off of separate clocks.

This commit also removes the native_mem's ability to replace
dc_cycle_stamp_priv_ with its own variable.  I intend to bring this back
eventually because removing it has made WashingtonDC a bit slower.