Turn dc_mem into a static variable
I don't like how I have to use the memory-map region ID system to treat memory
as a special case within the JIT's x86_64 backend (for optimization reasons)
but the only other solution is involved greatly expanding the JIT IL so that
code can be implemented as IL code in MemoryMap.c and memory.c.  Someday I will
get that done, but for now this is as modular as it gets.