ARM7: don't call next_inst after a pipeline reset
The reason for this is that the pipeline reset won't fetch what should be the
next instruction.  So it'll increment the PC to the instruction that goes after
the instruct that the ARM just jumped to, and then that will end up being the
first instruction to be fetched and what *should* have been the first
instruction to get fetched won't get fetched at all.

With this fix, the PC the ARM just jumped to becomes the next instruction to
execute and then it will call next_inst after it finishes (unless if also
overwrites the PC).

This fix does not change the behavior of B and BL because they never called
next_inst to begin with.