ARM7: SWI instruction
Daytona USA doesn't use this instruction as far as I know, but it does encode a
few SWI instructions with the NV (never) condition code.  WashingtonDC's ARM7
decoder will panic as soon as it sees something it doesn't recognize regardless
of whether the conditional would have worked or not, so I need to have this
instruction implemented.

With this commit Daytona USA boots and gets in-game on the
snickerbockers/arm7_yet_again branch.