pvr2: don't support interlacing for writing
I don't know how this is supposed to work, tbh.  There are two start-of-field
registers for writing, but IDK how to tell if interlacing is enabled or not.
Reading the SPG control register is not correct because that is for reads, not
writes.

KallistiOS programs seem to use interlacing for reading, but then the start of
field 2 at write-time is 0, so it looks like the start of frame 2 is supposed
to be ignored at write-time?  IDK, I'll keep an eye on this shit in the future.

This commit fixes all the KallistiOS programs that stopped working in commit
1975165e29e1b0721da65df35cb51423869d2609.