MemoryMap: allow unmapped reads/writes for the ARM7
Evolution: The World of Sacred Device is doing this on bootup because it tries
to access an uninitialized pointer.  I'm pretty sure this is actually what
happens on real hardware as well because I see other emulators doing the same
thing.

For the record, I made a hardware test and confirmed that when it reads from
unmapped memory, the value is all zeroes.