Don't use fancy bit-twiddling tricks to implement memory maps
For the record, I do actually think that bit-twiddling is the optimal way to
implement this and it is completely feasible, but the old implementation wasn't
working so it's going in the trash and I'm replacing it with a more naive
implementation.  Eventually I'll redo it the right way and it will be faster
and it won't incorrectly map address ranges.