move types.h outside of the sh4 directory
It's not sh4-specific, and several other components are including it anyways.

I'm actually thinking about getting rid of it since reg32_t, addr32_t and
uint32_t are all used interchangeably anyways, but I'll save that decision for
another day.

With this commit, there are no more dependencies on SH4 code in the JIT code.