sh4: fix unaligned memory reads in sh4_inst_binary_fitrv_mxtrx_fv
I have a laptop running gcc version 6.3.0-18+deb9u1.  builds made on this laptop
crash in this instruction handler due to an unaligned memory read (the pointer
casts).

ANYWAYS, this doesn't actually fix the problem because that specific vintage of
gcc appears to have a bug in which it converts potentially non-aligned memcpy
accesses into x86 sse movaps instructions, meaning that this commit is actually
moot.  However, I'm committing it anyways because the old implementation was
technically inviting undefined behavior.

For the record, my desktop is running gcc 8.2.1 and this bug does not exist
there.  With the new commit gcc 8.2.1 actually loads each float one at a time
as it should, so this bug has apparently already been fixed in gcc.