The beginning of WashingtonDC's configuration file
For now all it lets you do is set the background color (for when the window
resultion doesn't match the virtual Dreamcast's resolution).  In the future it
will include more useful settings such as controller button bindings, firmware
image paths, etc.

The configuration file is currently optional.  It should be a text file named
'wash.cfg' located in the current directory.  In the future it will be read from
a permanent location somewhere in the user's home directory.

To set the background color, enter the following line into wash.cfg:

bgcolor 3d77c0

3d77c0 should be replaced with the desired color, and it should consist of six
hexadecimal digits.