cfg: rename the bgcolor config node to ui.bgcolor
The idea is to  maybe someday introduce some sort of namespace or object
system.  For now that's not implemented and the dot is literally part of the
config node's name, but doing it this way ensures that config files are
future-proofed.  Hopefully this will be the only time I ever rename a config
node.