change sign conventions control bindings for analog axes
The direction of an analog axis (the +/- at the end) now only signifies whether
or not the axis is inverted.  All axes can return values less than 0 now.

The Dreamcast-side of the code is now what filters out negative values from the
analog stick half-axes.

The old convention forced the users to know that triggers have full range but
directions on analog sticks only have half range, which is a stupid thing to
expect users to know.

neutral sign (ie one not specified by user) is now the same as a + sign,
meaning that + is the default.