pvr2: warn when there's an inconsistency between read and write fb dimensions
This is not an easy problem to resolve with OpenGL without ditching
render-to-texture and CPU framebuffer access.  Dreamcast games see the
framebuffer as a simple block of memory in video RAM even when they use teh GPU,
but I have to layer that on top of OpenGL textures.

I DON'T EVEN KNOW IF I'M ALWAYS GETTING THE ASPECT RATION RIGHT OR IF IT'S OFF
BY A FEW PIXELS, THIS IS DRIVING ME INSANE!