AICA: print channel sample frequencies
Also dump channels even when they're not 16-bit.  The one format that the
corlett doc is unused (no. 3 iirc) is actually used by the spiral-swirl
animation.  IDK what format it is, but you can sorta hear the right sound when
it's decoded as 16-bit signed so that can't be too far off.