ui: don't show the texture cache window by default
this is because it has poor performance.  It fetches every texture from
emulated video memory and converts it to an OpenGL-compatible format every
frame.  Obviously it would be faster and also more accurate to use the version
which the texture cache has already uploaded to OpenGL, but for now I do it this
way because I still haven't moved the gfx backend into the frontend code.