JIT profiler
this gives the profiler all information it needs to track blocks based on access
counts (but not on execution time).

This system is enabled/disabled at compile-time with the -DJIT_PROFILE=On/Off
switch in cmake.

One limitation is that this does not handle cache invalidations correctly;
if all code blocks are destroyed by a cache invalidate, then any code-blocks
created after the invalidate which are identical to code-blocks created before
the invalidate will be seen as new separate code blocks unless they were already
in the high score.