jit: add an optimization pass to delete "dead loads" which have no effect
ie cases where it loads a value into a slot then overwrites that value without
ever reading from it.

The reason this happens is that whenever the sh4 frontend access an SH4
register for the first time in a given code block, the frontend will
uncoditionally read that register in from the sh4 state as if the register was
going to be read from.

This could also have been fixed in the sh4 frontend code by only loading from
the sh4 state when the register needs to be read from, but doing it as a JIT IL
optimization pass will be simpler and less error-prone since there are less
opcodes to consider.  Doing it as a JIT IL pass will also mean that I don't
have to do this again when I'm writing the ARM7 frontend.