jit: x86_64: reduce register shuffling for jump addresses and hashes
We do this by making the register allocator choose the jump address register
and the jump hash register for slots which will host the jump address and jump
hash respectively.  Previously the behavior was to not give these slots special
consideration, and that would lead to those values being loaded into other
registers before being moved to the correct registers.