GD-ROM: clean up ambiguities between session and region
Previously I believed that GD-ROM discs had two sessions, and that the second
session was the high-density region.  Based on recent hardware testing, I now
understand that GD-ROM discs only have one session.  The high density region
partitions the disk but the hardware still considers it to be part of the first
session.