GD-ROM: fix DMA transfers when the guest asks for less data than is buffered
The pre-made IP.BIN bootstrap that most homebrews and scene cracks are using
does this.  AFAIK, official Dreamcast software always asks for exactly as much
data as is buffered; that is why we've never seen this bug before.

With this commit, MIL-CD homebrew games (the ones that I tested, that is) can
boot and get in-game.