only specify winsock as a libevent prereq, not a WashingtonDC prereq
if we make it a prereq for both, then CMake will fuck up the link order
on MinGW.

Also don't use find_library to get the linker flags.  The problem with
find_library is that it will pick up on MSVC's version of the winsock
lib, and that's not compatible with MinGW's headers.

with this commit, the windows build FINALLY works without disabling libevent.