clean up the way g2 dma channels are implemented.
For the most part this shouldn't change any behavior.  One behavior that *does*
get changed is that DMA transfers on the E1, E2, and DD channels will raise an
ERROR_UNIMPLEMENTED now; previously attempts to begin DMA transfers on those
channels would be ignored.  This new behavior is preferable to the old behavior
because the old behavior wasn't performing the DMA transfer anyways.