jit: don't use movabs to load destination address when storing a float register
This greatly cuts down the number of movabs instructions emitted per code block,
but does not completely eliminate it.

Eventually I want to emit it once when we enter the jit, and then have every
code block offset from the value that was loaded so that we never need to
emit a movabs instruction ever again.