ui: add a memory-dumper option to the menu
This adds a new git submodule for the file browser and also necessitates
upgrading to C++17 for the washingtondc frontend.  Since the core-library is not
impatcted by this, I have no problems with forcing people to deal with C++17.

There's also an issue wherein you can't change the directory in the memory dump
file browswer and that is annoying, but I'll get to it later.