make aspect ratios work the same way they do on analog TV
This means that the output display has a fixed aspect-ratio of 4:3 regardless
of what the resolution of the framebuffer is.  Excess scanlines above 480 don't
get displayed, and the horizontal resolution isn't always stretched to whatever
it needs to be to make the aspect ratio 4:3.

This commit fixes 320x240 games like bangai-o that were getting stretched out
vertically.