pvr2: only access texture memory via special functions.
we no longer reference the internal storage of texture memory from outside of
pvr2_tex_mem.c/pvr2_tex_mem.h.

The purpose of this commit is to make way for unifying the backing of the
32-bit and 64-bit texture memory paths together like they are on real hardware.
Only then will I be able to boot that one Legend of Zelda fangame.