Implement framebuffer and texture synchronization across 32-bit and 64-bit access modes.
With this commit and the SH4 ORA write workaround, the Zelda: Return of the
Hylian homebrew shows graphics.  There is some corruption which appears to be
caused by an incorrect linestride somewhere, but that appears to be a result of
a pre-existing bug which is not related to this commit.