gdrom: undo effects of commit fb24bd6756c441666f7e98610e6f71ae218e17c0
this commit was breaking sonic adventure, and I don't think it was solving any
real problems anyways.  In the future, I need to come up with a better
heuristic for this latency.

With this commit, Sonic Adventure's attract mode can run indefinitely without
failure.

This commit resolves issue #82 on github.