SH4 Interrupt controller fix
previously the interrupt controller had a bug wherein it not re-raise an
interrupt if an IRQ handler exited without clearing it.  This was causing
several games to get hung on missed interrupts.

This commit fixes that by storing interrupt state in the sources of the
interrupts instead of in the sh4 itself.

As a result of this commit, single-steps in the debugger now skip over delay
slots.  You can't set a hardware breakpoint to a delay slot now, either.

Games that this commit fixes and allows to go in-game include:

Ecco the Dolphin
Star Wars Episode I Racer
Toy Story 2
Puyo Puyo 4

There are also several games which still don't go in-game but do progress
further before getting stuck on unimplemented functionality or unrelated bugs:

Zero Gunner 2
Tennis 2K2
Marvel vs Capcom 2
Ghetto Pong (homebrew)

This commit resolves issue #80 on github.