sh4: partially implement UTLB/ITLB MMIO access via address and data arrays
This  does not implement actual MMU functionality, it just lets programs read
and write to the TLB when it's not enabled and feel like they're doing
something.

There are some games which don't use the MMU but nevertheless still clear the
the valid bits on every entry.

With this commit, the emulated Genesis games in SEGA Smash Pack are now
in-game.  Virtua Cop and SEGA Swirl are still not in-game.  Also, Quake III
Arena shows its intro FMV although it still doesn't get in-game or in-menu.