sh4: wrapper functions for operand memory access (interpreter only)
the point of this is to provide a common point where MMU translations can be
inserted.  This is only implemented for data transfer operations in the
interpreter.  It is not implemented for the JIT, nor is it implemented for
instruction fetch, nor is it implemented in the DMAC or the store queues.

This will most likely require further modification in the future since an MMU
exception needs to be able to prematurely end the current opcode and raise an
exception.