SH4: MMU data TLB miss exceptions for 16-bit write
With this commit, the basic/mmu/nullptr example in the KallistiOS source works
on WashingtonDC.  It will crash the emulator due to an unimplemented register
write, but that happens after it catches the null pointer access and prints the
correct error message to the serial port.

Currently the exceptions are only implemented for 16-bit writes; all other
operations that result in a TLB exception will raise an error.  I'm holding off
on those until I have a testcase.

Also, TLB mappings still aren't implemented for instruction fetch at all, so
there's that.