sh4: add in MMU support to the storage queues
with this commit most of the non-Windows games which use the MMU are now
in-game in interpreter mode with -DENABLE_MMU=On.

Two exceptions to this are Sonic Shuffle and Neon Genesis Evangelion Rei Ayanami
Raising Project.  Sonic Shuffle hangs at startup, while Evangelion crashes doing
a couple weird things that are probably safe to ignore but I want to first
confirm on hardware.