hollywood: add new GD-DMA interrupt and allow it to be injected via the debug UI
This interrupt is not yet integrated into the GD-ROM code because it requires
more infrastructure.  This interrupt should be raised upon completion of a
GD-ROM DMA transfer.  The GD-ROM EXT interrupt should only be raised once all
data has been transferred from the ATAPI READ command; it is NOT to be raised
at the end of every DMA transfer (which is the current behavior).  This
interrupt is important to Windows CE and apparently not much else since I've
been able to get this far without even knowing it exists.