PVR2: List continuation processing
list continuation processing for non-windows games is still pretty wonky.  Part
of the problem is that I really should do some sort of low-level display list
tracking so I can be absolutey sure of what's going on, but the real issue is
just that I don't have good docs on how this should work.

I see a lot of games to a continuation immediately after closing a polygon
group, which makes me wonder if maybe the continuation invalidates the previous
close?  That doesn't completely make sense (what about the interrupt?) so I
haven't implemented it that way, but there's definitely SOMETHING weird like
that going on.