PVR2: better display list HLE
Instead of immediately rendering everything that comes in over the TAFIFO,
we now cache display lists and replay them when the STARTRENDER register is
written to.  This is still an HLE approach because we don't actually write
the object lists out to VRAM, but it's much closer to actual hardware than
what we had before.

This fixes the erroneous polygons visible on first frame of bootup
(issue #10 on github), the missing dev logos in Namco Museum, and many games
which were previously failing due to incomprehensible TAFIFO traffic (especially
the ones which happened when STARTRENDER was issued during group submission - those
games were actually keeping multiple frames in flight!).