sh4: don't use bitfields for utlb and itlb entries
the sz member is an enum, and there is apparently some
implementation-defined behavior surrounding that case.
When reading the sz member, gcc was emitting shr whereas
msvc was emitting sar.  This ironically led to Windows games
crashing the emulator on Windows since sz would be read as a negative
number when built with MSVC.

This might have negative performance implications, but the MMU
implementation never performed well anyways.