gfx: convert w to z in the backend
this makes it more versatile and gives me freedom to write better
debug tools.

Unfortunately, there is no easy way to do this for the clip_min/clip_max, which
is unfortunate because it leads to an inconsistent-design.  implementing this
would require the backend to buffer an entire frame of il commands, which is what
the frontend is already doing.