gfx: make the gfx backend aware of the PowerVR2 pixel clip rectangle
The old way of doing things was to handle the clip rectangle entirely
in the PowerVR2 code.  The problem with this was that the gfx backend
didn't know that the linestride was supposed to be greater than the
line width.  This was causing soft_gfx to output the wrong linestride in
Star Wars Episode I Racer.

The OpenGL renderer probably also had this exact same bug, but since the
framebuffer stayed entirely in OpenGL and was only ever copied to the display
window via texture mapping, we never saw the effects of having the wrong
linestride.  If it had been sync'd back to emulated texture memory then it
would have been incorrect.