gfx: make the w-component part of the vertex position
previously, verts only had 3 components for the position and the
w-component was hardcoded to 1 in the vertex shader.  w is still
hardcoded to 1, but it's done in the pvr2 code instead of the
vertex shader.

I'm of the opinion that this makes things simpler, plus the graphics
driver was probably padding this out to a 4-component vector anyways
to make it a power-of-two.