gfxgl4: eliminate the perspective divide
We preserve the original w-coordinate in each vertex instead of trying to
un-project them.  This fixes z-fighting BIGLY because we're no longer going
through an additional two perspective divisions (on top of the one the game
probably did).

After vertx shader, the w-coordinate of each projected vertex is now hardcoded
to 1.  This necessitates that the fragment and vertex shaders add in the
missing factors to make the per-vertex attributes interpolate
perspective-correct.

This commit does problems in the menus in Sonic Adventure; that's a classic
case of z-fighting where there are polygons with extremely small z-gaps
between them - there's still some imprecision in the pipeline and that's enough
to screw up Sonic Adventure.  This wasn't a problem when we were doing 1/z since
that turns very small gaps into very large gaps.

SoulCalibur also has issues because it sometimes renders polygons with extremely
large Z coords, which pushes our farplane out too far because we determine near
and far planes based on maximum and minimum Z values.