gfxgl4: use glClipControl and DEPTH_COMPONENT_32F to get mostly-good depth buffer precision
DEPTH_COMPONNET32F gives us similar precision to what the PVR2 has (in fact,
the only problem is that depth values are not unbounded, if they were this would
be perfect).

glClipControl lets us change the depth range from -1..1 to 0..1.  I'm not sure why
this is necessary, but i guess it gives us just barely enough precision to fix a lot
of things that were broken.

With this commit, Z fighting is greatly reduced in gfxgl4 albeit not totally solved.
I still need to come up with a solution that'll work in gfxgl3 as well as Vulkan
implementations that might not support 32-bit floating point depth buffer.