soft_gfx: refactor vertex attributes in terms of linear equations rather than geometry
i was doing this initially as a way to get rid of all the abstractions in soft_gfx's
rendering code so that I could get a better idea of where the floating-point
inaccuracies in the texture coordinates are coming from (see issue #91 on github), but
somehow the refactor itself reduced the inaccuracies.  Not entirely sure why, but it's
probably related to the way that I scale things by (x_pos - bbox[0]) and (y_pos - bbox[1])
instead of just x_pos and y_pos.

This doesn't totally resolve the inaccuracies, there's still more work to be done.