preliminary support for PuruPuru vibration emulation
it's only in the emulation code though, there's nothing hooking it up to the
frontend to actually vibrate the controller.  This is due in part to GLFW's
lack of support for force-feedback, and in part due to my lack of information
on how to decode the 4-byte value in the maplebus setcond packets that tells
the PuruPuru how to vibrate.

So for now it's disabled by default and if you do enable it then all it does is
printf to stdout when one of the controllers should be vibrating.