GD-ROM: let the ATA IDENTIFY command on through and treat it as a NOP
DreamShell uses this, I think it's poking about for systems with HDD mods.

AFAIK this is specific to normal IDE HDDs and the correct response for an
ATAPI device is to ignore it but I could be wrong on that.  This will be
trivial to verify if/when I finally have access to a real Dreamcast again.