pvr2: use one vertex array for all vertices in a given frame
this gives a tidy performance boost because overwriting the same vertex buffer
every time we issued a draw command was causing OpenGL to stall while it
finishes using the vertex data from the previous draw command.

Implementing this necessitated splitting the gfx draw command into two separate
commands: one that uploads vertex data and one that renders a subset of that
vertex data.

This also introduces a new rendering bug in Sonic Adventure wherein the dialogue
text will flicker a bit and I'm not sure why that but it defintely needs to be
fixed before this can merge to master.  It's pretty noticeable in the opening
cutscene of Sonic's story where he sees the cops chasing after Chaos 0.