transform texture coordinates for stride textures using a matrix in the rend_param
the matrix transformation happens in the gfx backend so that it can be offloaded
to OpenGL.

This eliminates a branch, a multiplication and a division that were all being
done per-vertex in the PVR2 emulation code.  Now we just have a branch and a
division that only get executed per header.