PARTIAL COMMIT - DO NOT MERGE TO MASTER
make all the jit_* functions return the index of the il instruction they just
added.

I want to gradually refactor the jit IL to do away with the concept of "slots"
and just have each instruction reference its inputs using the indices of the IL
instructions which produced those inputs.  This way the list of il instructions
will effectively be a dependency graph, so we can make more intelligent
decisions about register allocations and optimizations (this way the x86_64 can
do cool stuff like copy a value to a a different register and be aware that the
same value is in two places at once, which can help with stuff like the register
shuffling around function calls).

This ought to be possible since none of the IL insts has more than one output
anyways, but it's also a pretty lofty goal because there's a lot of code that
needs to be re-written so i don't think i have time to finish it before the
doctors cut up my leg.  also it probably won't lead to gains as big as i'd get
from finally implementing fastmem, or giving the ARM7 its own jit frontend.  but
it would be pretty sick.