gdrom: don't raise an error for out-of-bounds GD-ROM reads.
2K sports games are doing 32767-sector reads that go past the end of the final
track.  I'm still not sure if this is the correct behavior or if there is some
other error which is causing it to issue the abnormal read commands.

with this commit, NBA 2K is now in-game.  NFL 2K is not in-game yet because it
hangs at a blank screen.

see issue #97 on github.