add libchdr to the source tree as a git submodule
currently it points to an unofficial branch on my personal
fork on github.  This branch has been submitted upsteam in
a pull request to libchdr, but it has not been approved yet
so we'll have to use my fork for now.  hopefully it will
be approved and merged and then we will be able to point
the submodule at the upstream libchdr master branch.