replace libpng with nothings' stb_image_write.h
the reason for this is that im going to be adding libchdr
to WashingtonDC for loading compressed CHD images, but it
brings in a version of zlib that conflicts with the version
of zlib that we have as a pre-requisite.  Easiest solution
to this problem is to replace libpng with a different png
library that doesn't have zlib as a pre-requisite.

this also means that it would be possible to let users
choose a different format such as jpg because stb_image_write
supports more than png.  However, this option is not implemented.

source: https://github.com/nothings/stb
        HEAD af1a5bc352164740c1cc1354942b1c6b72eacb8a