gd-rom: implement late buffering of data
some of the 2k sports games like to issue extremely large read commands and
then only read a small fraction of that data from the drive controller.  This
causes absolutely terrible performance (less than one frame per second) because
the emulator is spending all of its time reading data from disc that will never
actually be used by the game.

To resolve this, the gd-rom drive emulation now reads data from the host
filesystem when it is either transferred by DMA or PIO into emulated memory.
This way only the data which is actually going to be used is read from the host
filesystem.